.history/requirements_20250311162822.txt
# Python dependencies
pip>=25.0.1
setuptools>=69.0.3
wheel>=0.42.0
MetaTrader5>=5.0.45
pandas>=2.2.0
numpy>=1.26.0
python-telegram-bot>=20.8
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.5
aiohttp>=3.9.0
ta-lib>=0.4.28
jinja2>=3.1.3
pydantic>=2.6.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4 